perf(store): reformat store file
Ref to issue #111, I would like to suggest a way to manage the store file more easier, welcome to
take ref for this commit and do a better version by you guys

BREAKING CHANGE: Reformat the Store

re #111
