+++ src/store/cacheUpdateAt/index.js
@@ -1,0 +1,13 @@
+export default {
+  state: {},
+  mutations: {
+    storeCacheUpdateAt: (state, d) => {
+      state = Object.assign(state, d);
+    }
+  },
+  getters: {
+    cacheUpdateAt: (state) => (name) => {
+      return state[name]
+    }
+  }
+};
