+++ src/store/dataSource/index.js
@@ -1,0 +1,8 @@
+export default {
+  state: 'global',
+  mutations: {
+    switchDataSource: (state, value) => {
+      state = value;
+    }
+  }
+};
